+++ packages/extension-polkagate/src/fullscreen/governance/index.tsx
@@ -77,7 +77,7 @@ export default function Governance (): React.ReactElement {
     fetchJson();
   }, []);
 
-  const referendaTrackId = tracks?.find((t) => String(t[1].name) === selectedSubMenu.toLowerCase().replace(' ', '_'))?.[0]?.toNumber()!;
+  const referendaTrackId = tracks?.find((t) => String(t[1].name) === selectedSubMenu.toLowerCase().replaceAll(' ', '_'))?.[0]?.toNumber();
   // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
   const currentTrack = useMemo(() => {
     if (!tracks && !fellowshipTracks) {
@@ -86,7 +86,7 @@ export default function Governance (): React.ReactElement {
 
     // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     return (tracks || []).concat(fellowshipTracks || []).find((t) =>
-      String(t[1].name) === selectedSubMenu.toLowerCase().replace(' ', '_') ||
+      String(t[1].name) === selectedSubMenu.toLowerCase().replaceAll(' ', '_') ||
       String(t[1].name) === selectedSubMenu.toLowerCase() // fellowship tracks have no underscore!
     );
   }, [fellowshipTracks, selectedSubMenu, tracks]);
@@ -212,9 +212,6 @@ export default function Governance (): React.ReactElement {
       return;
     }
 
-    fetchRef(_key).then(() => setIsFetching(false)).catch(console.error);
-    setReferenda(refsContext.refs?.[_key]);
-
     async function fetchRef (key: string) {
       if (!chainName) {
         return;
@@ -227,7 +224,7 @@ export default function Governance (): React.ReactElement {
       // Reset referenda list on menu change
       if (isSubMenuChanged || isTopMenuChanged) {
         // setReferenda(undefined);
-        // setFilteredReferenda(undefined);
+        setFilteredReferenda(undefined);
         list = [];
         pageTrackRef.current.subMenu = selectedSubMenu; // Update the ref with new values
         pageTrackRef.current.page = 1;
@@ -293,6 +290,9 @@ export default function Governance (): React.ReactElement {
 
       handleSettingReferenda(key, concatenated);
     }
+
+    fetchRef(_key).then(() => setIsFetching(false)).catch(console.error);
+    setReferenda(refsContext.refs?.[_key]);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [addFellowshipOriginsFromSb, chainName, fellowshipTracks, _key, getMore, isSubMenuChanged, isTopMenuChanged, referendaTrackId, selectedSubMenu, topMenu, tracks]);
 
@@ -381,13 +381,11 @@ export default function Governance (): React.ReactElement {
             decidingCounts={decidingCounts}
             menuOpen={menuOpen}
             setMenuOpen={setMenuOpen}
-            // @ts-ignore
             setSelectedSubMenu={setSelectedSubMenu}
           />
           <Container disableGutters sx={{ maxWidth: 'inherit' }}>
             <Bread
               address={address}
-              // @ts-ignore
               setSelectedSubMenu={setSelectedSubMenu}
               subMenu={selectedSubMenu}
               // @ts-ignore